+++ R/tbl-df.r
@@ -7,13 +7,16 @@
 #'
 #' @section Methods:
 #'
-#' \code{tbl_df} implements two important base methods:
+#' \code{tbl_df} implements four important base methods:
 #'
 #' \describe{
 #' \item{print}{Only prints the first 10 rows, and the columns that fit on
 #'   screen}
 #' \item{\code{[}}{Never simplifies (drops), so always returns data.frame}
+#' \item{\code{[[}, \code{$}}{Calls \code{\link{.subset2}} directly,
+#'   so is considerably faster.}
 #' }
+#'
 #'
 #' @export
 #' @param data a data frame
@@ -119,7 +122,17 @@ print.tbl_df <- function(x, ..., n = NULL, width = NULL) {
 }
 
 #' @export
-`[.tbl_df` <- function (x, i, j, drop = FALSE) {
+`[[.tbl_df` <- function(x, i) {
+  .subset2(x, i)
+}
+
+#' @export
+`$.tbl_df` <- function(x, i) {
+  .subset2(x, i)
+}
+
+#' @export
+`[.tbl_df` <- function(x, i, j, drop = FALSE) {
   if (missing(i) && missing(j)) return(x)
   if (drop) warning("drop ignored", call. = FALSE)